requirements.txt
fastapi==0.120.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.1
pydantic==2.12.3
pydantic-settings==2.11.0
python-slugify==8.0.1
python-dotenv==1.0.0
email-validator
psyocopg2
psycopg2-binary
greenlet